add logging to locktests too
