+++ Consul.Test/TestOutputConverter.cs
@@ -1,0 +1,69 @@
+// -----------------------------------------------------------------------
+//  <copyright file="TestOutputConverter.cs" company="G-Research Limited">
+//    Copyright 2020 G-Research Limited
+//
+//    Licensed under the Apache License, Version 2.0 (the "License"),
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+//  </copyright>
+// -----------------------------------------------------------------------
+
+using System;
+using System.IO;
+using System.Text;
+using Xunit.Abstractions;
+
+namespace Consul.Test;
+
+public class TestOutputConverter : TextWriter
+{
+    ITestOutputHelper _output;
+
+    public TestOutputConverter(ITestOutputHelper output)
+    {
+        _output = output;
+    }
+
+    public override Encoding Encoding
+    {
+        get { return Encoding.Default; }
+    }
+
+    public override void WriteLine(string message)
+    {
+        try
+        {
+            _output.WriteLine(DateTime.Now.ToString("HH:mm:ss.ffff") + ": " + message);
+        }
+        catch (Exception)
+        {
+            // ignored
+        }
+    }
+
+    public override void WriteLine(string format, params object[] args)
+    {
+        try
+        {
+            _output.WriteLine(DateTime.Now.ToString("HH:mm:ss.ffff") + ": " + format, args);
+        }
+        catch (Exception)
+        {
+            // ignored
+        }
+    }
+
+    public override void Write(char value)
+    {
+        throw new NotSupportedException(
+            "This text writer only supports WriteLine(string) and WriteLine(string, params object[]).");
+    }
+}
